requirements.txt
streamlit>=1.20
openai>=1.59.0
streamlit-authenticator
python-dotenv
